refactor: moved music-player.api.ts into api folder
