+++ src/components/LoginForm/LoginForm.tsx
@@ -9,7 +9,7 @@ import { Button } from '../Button/Button'
 import './LoginForm.css'
 import { InputField } from '../InputField/InputField'
 import { useNavigate } from 'react-router-dom'
-import { ISignupUser, useUserSignupMutation, useUserTokenMutation } from '../../slices/music-player.api'
+import { ISignupUser, useUserSignupMutation, useUserTokenMutation } from '../../api/music-player.api'
 import { useCookies } from 'react-cookie'
 import { useAppDispatch } from '../../hooks/appHooks'
 import { useLogout } from '../../hooks/userHooks'
