+++ src/slices/userSlice.ts
@@ -1,6 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { IUser } from '../models';
-import { musicPlayerApi } from './music-player.api';
+import { musicPlayerApi } from '../api/music-player.api';
 import { RootState } from '../store';
 
 const initialState: IUser = {
