+++ src/store.ts
@@ -4,7 +4,7 @@ import userReducer from './slices/userSlice';
 import filterReducer from './components/Filter/FilterSlice';
 import trackReducer from './slices/trackSlice';
 import playerReducer from './slices/playerSlice';
-import { musicPlayerApi } from './slices/music-player.api';
+import { musicPlayerApi } from './api/music-player.api';
 
 export const store = configureStore({
   reducer: {
